requirements.txt
asgiref==3.3.1
colorama==0.4.4
Django==3.1.3
Pillow==8.0.1
pytz==2020.4
qrcode==6.1
six==1.15.0
sqlparse==0.4.1
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
django-dash==0.5.6
django-otp==1.0.2
reportlab-qrcode==0.1.4
reportlab==3.5.66
django-filter==2.4.0
selenium==3.141.0
chromium-chromedriver==90.0.4430.72
